scalability framework: rename benchmark_result.py to avoid name clash
